requirements.txt
uvicorn~=0.30.3
fastapi~=0.111.1
aiogram~=3.4.1
python-dotenv~=1.0.1
starlette~=0.37.2